Refactor image sequence generation into separate file
